+++ tree-sitter-harn/queries/highlights.scm
@@ -14,14 +14,34 @@
 "try" @keyword.exception
 "catch" @keyword.exception
 "throw" @keyword.exception
+"finally" @keyword.exception
 "return" @keyword.return
 "import" @keyword.import
 "fn" @keyword.function
 "spawn" @keyword
 "parallel" @keyword
 "parallel_map" @keyword
+"parallel_settle" @keyword
 "type" @keyword
 "pub" @keyword
+"enum" @keyword
+"struct" @keyword
+"impl" @keyword
+"interface" @keyword
+"where" @keyword
+"break" @keyword
+"continue" @keyword
+"yield" @keyword
+"deadline" @keyword
+"guard" @keyword
+"mutex" @keyword
+"select" @keyword
+"from" @keyword
+"timeout" @keyword
+"default" @keyword
+"not" @keyword.operator
+"upto" @keyword.operator
+"thru" @keyword.operator
 
 ; Literals
 (true) @boolean
@@ -29,6 +49,7 @@
 (nil) @constant.builtin
 (integer_literal) @number
 (float_literal) @number.float
+(duration_literal) @number
 (string_literal) @string
 (interpolated_string) @string
 (string_content) @string
@@ -68,6 +89,43 @@
 (type_declaration
   name: (identifier) @type.definition)
 
+; Enum declarations
+(enum_declaration
+  name: (identifier) @type)
+
+(enum_variant
+  name: (identifier) @constant)
+
+; Struct declarations
+(struct_declaration
+  name: (identifier) @type)
+
+(struct_field
+  name: (identifier) @property)
+
+; Impl blocks
+(impl_block
+  type_name: (identifier) @type)
+
+; Interface declarations
+(interface_declaration
+  name: (identifier) @type)
+
+(interface_method
+  name: (identifier) @function)
+
+; Generic params
+(generic_params
+  (identifier) @type)
+
+; Where clause
+(where_clause
+  (identifier) @type)
+
+; Select
+(select_case
+  variable: (identifier) @variable)
+
 ; Type annotations
 (type_annotation
   (identifier) @type)
@@ -75,6 +133,10 @@
 ; Shape type fields
 (shape_field
   name: (identifier) @property)
+
+; Dict entry keys
+(dict_entry
+  key: (identifier) @property)
 
 ; Operators
 "|>" @operator
@@ -91,8 +153,15 @@
 "-" @operator
 "*" @operator
 "/" @operator
+"%" @operator
 "!" @operator
 "=" @operator
+"+=" @operator
+"-=" @operator
+"*=" @operator
+"/=" @operator
+"%=" @operator
+"..." @operator
 "->" @punctuation.delimiter
 "?" @operator
 ":" @punctuation.delimiter